requirements.txt
pytest
playwright
requests
qase-pytest==4.2.0
python-dotenv
